refactor: move label functions under label.py and create basic tests
