+++ data_bridges_knots/client.py
@@ -4,12 +4,14 @@ import logging
 import os
 import time
 from datetime import date
+
 import data_bridges_client
 import numpy as np
 import pandas as pd
 import yaml
 from data_bridges_client.rest import ApiException
 from data_bridges_client.token import WfpApiToken
+
 from data_bridges_knots.helpers import get_adm0_code
 
 logname = "data_bridges_api_calls.log"
@@ -22,6 +24,7 @@ logging.basicConfig(
 )
 
 logger = logging.getLogger(__name__)
+
 
 def config_from_env() -> Dict:
     """Construct DataBridges configuration dictionary from environment variables.
@@ -288,7 +291,9 @@ class DataBridgesShapes:
         page = 0
         while total_items > max_item:
             page += 1
-            with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+            with data_bridges_client.ApiClient(
+                self._setup_configuration_and_authentication(self.config)
+            ) as api_client:
                 api_instance = data_bridges_client.MarketPricesApi(api_client)
                 env = self.env
 
@@ -355,7 +360,9 @@ class DataBridgesShapes:
         page = 0
         while total_items > max_item:
             page += 1
-            with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+            with data_bridges_client.ApiClient(
+                self._setup_configuration_and_authentication(self.config)
+            ) as api_client:
                 api_instance = data_bridges_client.CurrencyApi(api_client)
                 env = self.env
 
@@ -406,7 +413,9 @@ class DataBridgesShapes:
         Raises:
             ApiException: If there's an error calling the Food Security API
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             food_security_api_instance = data_bridges_client.FoodSecurityApi(api_client)
             env = self.env
 
@@ -458,7 +467,9 @@ class DataBridgesShapes:
         Returns:
             pandas.DataFrame: A DataFrame containing the retrieved commodity data.
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.CommoditiesApi(api_client)
             env = self.env
 
@@ -518,7 +529,9 @@ class DataBridgesShapes:
         Returns:
             pandas.DataFrame: A DataFrame containing the retrieved conversion factors.
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.CommodityUnitsApi(api_client)
             env = self.env
 
@@ -574,7 +587,9 @@ class DataBridgesShapes:
         Returns:
             pandas.DataFrame: A DataFrame containing the retrieved commodity units data.
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.CommodityUnitsApi(api_client)
             env = self.env
 
@@ -629,7 +644,9 @@ class DataBridgesShapes:
         Returns:
             pandas.DataFrame: A DataFrame containing the retrieved currency data.
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.CurrencyApi(api_client)
             env = self.env
 
@@ -680,7 +697,9 @@ class DataBridgesShapes:
         Returns:
             pandas.DataFrame: A DataFrame containing the retrieved exchange rate data.
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.CurrencyApi(api_client)
             env = self.env
 
@@ -724,7 +743,9 @@ class DataBridgesShapes:
         Returns:
             pandas.DataFrame: A DataFrame containing the retrieved economic indicator data.
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             # Create an instance of the API class
             api_instance = data_bridges_client.EconomicDataApi(api_client)
 
@@ -753,7 +774,9 @@ class DataBridgesShapes:
         adm0code = get_adm0_code(country_iso3)
 
         # Enter a context with an instance of the API client
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             # Create an instance of the API class
             api_instance = data_bridges_client.MarketsApi(api_client)
 
@@ -797,7 +820,9 @@ class DataBridgesShapes:
             ApiException: If there's an error accessing the Markets API
         """
         # Enter a context with an instance of the API client
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             # Create an instance of the API class
             api_instance = data_bridges_client.MarketsApi(api_client)
             format = "json"  # str | Output format: [JSON|CSV] Json is the default value (optional) (default to 'json')
@@ -843,7 +868,9 @@ class DataBridgesShapes:
 
         adm0code = get_adm0_code(country_iso3)
 
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.MarketsApi(api_client)
             local_names = False  # bool | If true the name of markets and regions will be localized if available (optional) (default to False)
 
@@ -891,7 +918,9 @@ class DataBridgesShapes:
         """
 
         adm0code = get_adm0_code(country_iso3)
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.MarketsApi(api_client)
             env = self.env
 
@@ -943,7 +972,9 @@ class DataBridgesShapes:
             ValueError: If data_type is not one of the allowed values
             ApiException: If there's an error accessing the GORP API
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             gorp_api_instance = data_bridges_client.GorpApi(api_client)
             env = self.env
 
@@ -1010,7 +1041,9 @@ class DataBridgesShapes:
 
         while total_items > max_item:
             page += 1
-            with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+            with data_bridges_client.ApiClient(
+                self._setup_configuration_and_authentication(self.config)
+            ) as api_client:
                 api_instance = data_bridges_client.IncubationApi(api_client)
                 env = self.env
 
@@ -1103,7 +1136,9 @@ class DataBridgesShapes:
 
         adm0code = get_adm0_code(country_iso3) if country_iso3 else None
 
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.IncubationApi(api_client)
             env = self.env
 
@@ -1149,7 +1184,9 @@ class DataBridgesShapes:
         Raises:
             ApiException: If there's an error accessing the API
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.IncubationApi(api_client)
             env = self.env
 
@@ -1222,7 +1259,9 @@ class DataBridgesShapes:
         Returns:
             bytes: The downloaded data as bytes.
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.IncubationApi(api_client)
             env = self.env
 
@@ -1251,7 +1290,9 @@ class DataBridgesShapes:
         Returns:
             bytes: The downloaded polygon files as bytes.
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.IncubationApi(api_client)
             env = self.env
 
@@ -1275,7 +1316,9 @@ class DataBridgesShapes:
         """
         Get a full dataset that includes all the fields included in the survey in addition to the core Market Functionality Index (MFI) fields by Survey ID.
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.SurveysApi(api_client)
             env = self.env
             try:
@@ -1301,7 +1344,9 @@ class DataBridgesShapes:
         """
         Retrieve Survey IDs, their corresponding XLS Form IDs, and Base XLS Form of all MFI surveys conducted in a country.
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.SurveysApi(api_client)
             env = self.env
 
@@ -1338,7 +1383,9 @@ class DataBridgesShapes:
         """
         Get MFI processed data in long format.
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.SurveysApi(api_client)
             env = self.env
 
@@ -1369,7 +1416,9 @@ class DataBridgesShapes:
         def get_rpme_base_data(
             self, survey_id=None, page: Optional[int] = 1, page_size=20
         ):
-            with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+            with data_bridges_client.ApiClient(
+                self._setup_configuration_and_authentication(self.config)
+            ) as api_client:
                 api_instance = data_bridges_client.RpmeApi(api_client)
                 env = self.env
 
@@ -1395,7 +1444,9 @@ class DataBridgesShapes:
             page: Optional[int] = 1,
             page_size=20,
         ):
-            with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+            with data_bridges_client.ApiClient(
+                self._setup_configuration_and_authentication(self.config)
+            ) as api_client:
                 api_instance = data_bridges_client.RpmeApi(api_client)
                 env = self.env
 
@@ -1427,7 +1478,9 @@ class DataBridgesShapes:
             market_id=None,
             adm0_code_dots="",
         ):
-            with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+            with data_bridges_client.ApiClient(
+                self._setup_configuration_and_authentication(self.config)
+            ) as api_client:
                 api_instance = data_bridges_client.RpmeApi(api_client)
                 env = self.env
 
@@ -1455,7 +1508,9 @@ class DataBridgesShapes:
         def get_rpme_surveys(
             self, adm0_code=0, page: Optional[int] = 1, start_date=None, end_date=None
         ):
-            with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+            with data_bridges_client.ApiClient(
+                self._setup_configuration_and_authentication(self.config)
+            ) as api_client:
                 api_instance = data_bridges_client.RpmeApi(api_client)
                 env = self.env
 
@@ -1479,7 +1534,9 @@ class DataBridgesShapes:
 
         # TODO: Get the scope and test these functions
         def get_rpme_variables(self, page: Optional[int] = 1):
-            with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+            with data_bridges_client.ApiClient(
+                self._setup_configuration_and_authentication(self.config)
+            ) as api_client:
                 api_instance = data_bridges_client.RpmeApi(api_client)
                 env = self.env
 
@@ -1499,7 +1556,9 @@ class DataBridgesShapes:
         def get_rpme_xls_forms(
             self, adm0_code=0, page: Optional[int] = 1, start_date=None, end_date=None
         ):
-            with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+            with data_bridges_client.ApiClient(
+                self._setup_configuration_and_authentication(self.config)
+            ) as api_client:
                 api_instance = data_bridges_client.RpmeApi(api_client)
                 env = self.env
 
@@ -1526,7 +1585,9 @@ class DataBridgesShapes:
     def get_mfi_xls_forms(
         self, adm0_code=0, page: Optional[int] = 1, start_date=None, end_date=None
     ):
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.XlsFormsApi(api_client)
             env = self.env
 
@@ -1574,7 +1635,9 @@ class DataBridgesShapes:
         Returns:
             pandas.DataFrame: DataFrame containing XLS Forms data
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.XlsFormsApi(api_client)
             env = self.env
 
@@ -1622,7 +1685,9 @@ class DataBridgesShapes:
         Returns:
             pandas.DataFrame: DataFrame containing MFI base survey data
         """
-        with data_bridges_client.ApiClient(self._setup_configuration_and_authentication(self.config)) as api_client:
+        with data_bridges_client.ApiClient(
+            self._setup_configuration_and_authentication(self.config)
+        ) as api_client:
             api_instance = data_bridges_client.SurveysApi(api_client)
             env = self.env
