+++ data_bridges_knots/helpers.py
@@ -1,86 +1,7 @@
 from typing import Dict, Optional
+
 import json
 from pathlib import Path
-import pandas as pd
-
-
-def get_value_labels(xlsform_df: pd.DataFrame, as_dataframe = True) -> Dict:
-    choiceList = pd.json_normalize(xlsform_df["choiceList"])
-    choiceList = choiceList.rename(columns={"name": "choice_name"})
-    choiceList = choiceList.join(xlsform_df["name"]).dropna()
-    choices = choiceList.explode("choices")
-
-    categories_dict = {}
-    for _, row in choices.iterrows():
-        name = row["name"]
-        choice = row["choices"]
-        if name in categories_dict:
-            categories_dict[name].update({(choice["name"]): choice["label"]})
-        else:
-            categories_dict[name] = {(choice["name"]): choice["label"]}
-
-    if as_dataframe = True:
-        return pd.DataFrame(categories_dict)
-    
-    return categories_dict
-
-
-
-def get_column_labels(xlsform_df: pd.DataFrame) -> Dict:
-    labels_dict = {}
-
-    for _, row in xlsform_df.iterrows():
-        name = row["name"]
-        label = row["label"]
-        if name in labels_dict:
-            labels_dict[name].update(label)
-        elif label == "":
-            labels_dict[name] = name
-        else:
-            labels_dict[name] = label
-    return labels_dict
-
-
-# Map values if int
-def map_value_labels(survey_df, xlsform_df):
-    survey_data = survey_df.convert_dtypes()
-    choiceList = pd.json_normalize(xlsform_df["choiceList"])
-    choiceList = choiceList.rename(columns={"name": "choice_name"})
-    choiceList = choiceList.join(xlsform_df["name"]).dropna()
-    choices = choiceList.explode("choices")
-
-    categories_dict = dict()
-    for _, row in choices.iterrows():
-        name = row["name"]
-        choice = row["choices"]
-        if name in categories_dict:
-            categories_dict[name].update({(choice["name"]): choice["label"]})
-        else:
-            categories_dict[name] = {(choice["name"]): choice["label"]}
-
-    # Map the categories to survey_data
-    survey_data_value_labels = survey_data.copy()
-    for col in survey_data_value_labels.columns:
-        if col in categories_dict:
-            category_dict = categories_dict[col]
-            survey_data_value_labels[col] = survey_data_value_labels[col].apply(
-                lambda x: category_dict.get(x, x)
-            )
-
-    return survey_data_value_labels
-
-
-def as_numeric(df, col_list):
-    for col in col_list:
-        try:
-            df[col] = (
-                pd.to_numeric(df[col], errors="ignore").fillna(9999).astype("int64")
-            )
-        except ValueError:
-            continue
-
-    return df
-
 
 # Cache return value from _load_country_codes() into global variable
 codes_mapping = None
@@ -138,3 +59,64 @@ def get_adm0_code(country_iso3: str) -> Optional[int]:
 
     codes = _load_country_codes()
     return codes.get(country_iso3.upper())
+
+
+def discover_surveys():
+    print("========================================\n")
+    print("DATABRIDGES SURVEY DISCOVERY\n")
+    print("========================================\n\n")
+
+    # Step 1: List all available surveys
+    print("========================================\n")
+    print("STEP 1: LIST AVAILABLE SURVEYS\n")
+    print("========================================\n")
+    print("Fetching household surveys...\n\n")
+
+    # Step 2: Search function
+    print("========================================\n")
+    print("STEP 2: SEARCH FOR YOUR SURVEY\n")
+    print("========================================\n\n")
+
+    # Step 3: Inspect a specific survey
+    print("========================================\n")
+    print("STEP 3: INSPECT A SPECIFIC SURVEY\n")
+    print("========================================\n")
+    print("Once you have a survey ID, inspect its data structure:\n\n")
+
+    # Step 4: Interactive exploration
+    print("========================================\n")
+    print("STEP 4: INTERACTIVE EXPLORATION\n")
+    print("========================================\n\n")
+
+    print("Available data:\n")
+    print("  available_surveys           # All surveys (if loaded)\n")
+    print("  current_survey_data         # Currently inspected survey data\n\n")
+
+    print("Available functions:\n")
+    print("  search_surveys('keyword')   # Search for surveys by name\n")
+    print("  inspect_survey(survey_id)   # Load and inspect a survey\n\n")
+
+    print("Example workflow:\n")
+    print("  1. search_surveys('BCM')                   # Find BCM surveys\n")
+    print("  2. inspect_survey(5387)                    # Inspect a specific survey\n")
+    print("  3. View(current_survey_data)               # Open in viewer\n")
+    print("  4. names(current_survey_data)              # See all columns\n")
+    print("  5. table(current_survey_data$column_name)  # Analyze a column\n\n")
+
+    print("Exploration commands:\n")
+    print("  View(current_survey_data)              # Open in viewer\n")
+    print("  names(current_survey_data)             # All column names\n")
+    print("  unique(current_survey_data$COLUMN)     # Unique values\n")
+    print("  table(current_survey_data$COLUMN)      # Count by printegory\n")
+    print("  str(current_survey_data)               # Data structure\n")
+    print("  summary(current_survey_data)           # Summary statistics\n\n")
+
+    print("========================================\n")
+    print("READY FOR DISCOVERY!\n")
+    print("========================================\n")
+    print("Complete workflow:\n")
+    print("  1. Browse available surveys above (Step 1)\n")
+    print("  2. Search for your survey (Step 2)\n")
+    print("  3. Inspect the survey structure (Step 3)\n")
+    print("  4. Explore the data (Step 4)\n")
+    print("  5. Create your exercise file using the information gathered\n\n")
