--- tests/test_fixture.py
@@ -1,3 +1,0 @@
-def test_sample_list_sum(sample_list):
-    """Test that the sample list sums to 6."""
-    assert sum(sample_list) == 6
