--- tests/test_load_stata.py
@@ -1,2 +1,0 @@
-def test_load_data():
-    pass
